gitserver: Remove data directory migration
This migration has existed pre-OSS so has run on all instances.
